compose: refactor quoting logic into QuotedMessagePresenter
