+++ k9mail/src/main/java/com/fsck/k9/activity/compose/RecipientPresenter.java
@@ -32,6 +32,7 @@ import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.Message.RecipientType;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mailstore.LocalMessage;
+import com.fsck.k9.message.PgpMessageBuilder;
 import com.fsck.k9.view.RecipientSelectView.Recipient;
 import org.openintents.openpgp.IOpenPgpService2;
 import org.openintents.openpgp.util.OpenPgpApi;
@@ -680,6 +681,11 @@ public class RecipientPresenter implements PermissionPingCallback {
             Log.e(K9.LOG_TAG, "obtained openpgpapi object, but service is not bound! inconsistent state?");
         }
         return new OpenPgpApi(context, openPgpServiceConnection.getService());
+    }
+
+    public void builderSetProperties(PgpMessageBuilder pgpBuilder) {
+        pgpBuilder.setOpenPgpApi(getOpenPgpApi());
+        pgpBuilder.setCryptoStatus(getCurrentCryptoStatus());
     }
 
     public enum CryptoProviderState {
