+++ k9mail/src/main/java/com/fsck/k9/ui/compose/QuotedMessagePresenter.java
@@ -1,0 +1,486 @@
+package com.fsck.k9.ui.compose;
+
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import android.content.res.Resources;
+import android.os.Bundle;
+import android.util.Log;
+
+import com.fsck.k9.Account;
+import com.fsck.k9.Account.MessageFormat;
+import com.fsck.k9.Account.QuoteStyle;
+import com.fsck.k9.K9;
+import com.fsck.k9.activity.MessageCompose;
+import com.fsck.k9.activity.MessageCompose.Action;
+import com.fsck.k9.helper.HtmlConverter;
+import com.fsck.k9.helper.QuotedMessageHelper;
+import com.fsck.k9.mail.Message;
+import com.fsck.k9.mail.MessagingException;
+import com.fsck.k9.mail.Part;
+import com.fsck.k9.mail.internet.MessageExtractor;
+import com.fsck.k9.mail.internet.MimeUtility;
+import com.fsck.k9.mailstore.LocalMessage;
+import com.fsck.k9.message.IdentityField;
+import com.fsck.k9.message.InsertableHtmlContent;
+import com.fsck.k9.message.MessageBuilder;
+import com.fsck.k9.message.QuotedTextMode;
+import com.fsck.k9.message.SimpleMessageFormat;
+import org.htmlcleaner.CleanerProperties;
+import org.htmlcleaner.HtmlCleaner;
+import org.htmlcleaner.SimpleHtmlSerializer;
+import org.htmlcleaner.TagNode;
+
+
+public class QuotedMessagePresenter {
+    private static final String STATE_KEY_HTML_QUOTE = "state:htmlQuote";
+    private static final String STATE_KEY_QUOTED_TEXT_MODE = "state:quotedTextShown";
+    private static final String STATE_KEY_QUOTED_TEXT_FORMAT = "state:quotedTextFormat";
+    private static final String STATE_KEY_FORCE_PLAIN_TEXT = "state:forcePlainText";
+
+    // Regexes to check for signature.
+    private static final Pattern DASH_SIGNATURE_PLAIN = Pattern.compile("\r\n-- \r\n.*", Pattern.DOTALL);
+    private static final Pattern DASH_SIGNATURE_HTML = Pattern.compile("(<br( /)?>|\r?\n)-- <br( /)?>", Pattern.CASE_INSENSITIVE);
+    private static final Pattern BLOCKQUOTE_START = Pattern.compile("<blockquote", Pattern.CASE_INSENSITIVE);
+    private static final Pattern BLOCKQUOTE_END = Pattern.compile("</blockquote>", Pattern.CASE_INSENSITIVE);
+
+
+    private final QuotedMessageMvpView view;
+    private final MessageCompose messageCompose;
+    private final Resources resources;
+
+    private QuotedTextMode quotedTextMode;
+    private QuoteStyle quoteStyle;
+    private boolean forcePlainText;
+
+    /**
+     * "Original" message body
+     *
+     * <p>
+     * The contents of this string will be used instead of the body of a referenced message when
+     * replying to or forwarding a message.<br>
+     * Right now this is only used when replying to a signed or encrypted message. It then contains
+     * the stripped/decrypted body of that message.
+     * </p>
+     * <p><strong>Note:</strong>
+     * When this field is not {@code null} we assume that the message we are composing right now
+     * should be encrypted.
+     * </p>
+     */
+    private String sourceMessageBody;
+
+    private SimpleMessageFormat quotedTextFormat;
+    private InsertableHtmlContent quotedHtmlContent;
+    private Account account;
+
+
+    public QuotedMessagePresenter(MessageCompose messageCompose, QuotedMessageMvpView quotedMessageMvpView,
+            Account account, String sourceMessageBody) {
+        this.messageCompose = messageCompose;
+        this.resources = messageCompose.getResources();
+        this.view = quotedMessageMvpView;
+        this.sourceMessageBody = sourceMessageBody;
+        onSwitchAccount(account);
+
+        quotedTextMode = QuotedTextMode.NONE;
+        quoteStyle = account.getQuoteStyle();
+
+        quotedMessageMvpView.setOnClickPresenter(this);
+    }
+
+    public void onSwitchAccount(Account account) {
+        this.account = account;
+    }
+
+    public void showOrHideQuotedText(QuotedTextMode mode) {
+        quotedTextMode = mode;
+        view.showOrHideQuotedText(mode, quotedTextFormat);
+    }
+
+    /**
+     * Build and populate the UI with the quoted message.
+     *
+     * @param showQuotedText
+     *         {@code true} if the quoted text should be shown, {@code false} otherwise.
+     */
+    public void populateUIWithQuotedMessage(Message sourceMessage, boolean showQuotedText, Action action)
+            throws MessagingException {
+        MessageFormat origMessageFormat = account.getMessageFormat();
+
+        if (forcePlainText || origMessageFormat == MessageFormat.TEXT) {
+            // Use plain text for the quoted message
+            quotedTextFormat = SimpleMessageFormat.TEXT;
+        } else if (origMessageFormat == MessageFormat.AUTO) {
+            // Figure out which message format to use for the quoted text by looking if the source
+            // message contains a text/html part. If it does, we use that.
+            quotedTextFormat =
+                    (MimeUtility.findFirstPartByMimeType(sourceMessage, "text/html") == null) ?
+                            SimpleMessageFormat.TEXT : SimpleMessageFormat.HTML;
+        } else {
+            quotedTextFormat = SimpleMessageFormat.HTML;
+        }
+
+        // TODO -- I am assuming that sourceMessageBody will always be a text part.  Is this a safe assumption?
+
+        // Handle the original message in the reply
+        // If we already have sourceMessageBody, use that.  It's pre-populated if we've got crypto going on.
+        String content = sourceMessageBody != null ? sourceMessageBody :
+                QuotedMessageHelper.getBodyTextFromMessage(sourceMessage, quotedTextFormat);
+
+        if (quotedTextFormat == SimpleMessageFormat.HTML) {
+            // Strip signature.
+            // closing tags such as </div>, </span>, </table>, </pre> will be cut off.
+            if (account.isStripSignature() &&
+                    (action == Action.REPLY || action == Action.REPLY_ALL)) {
+                Matcher dashSignatureHtml = DASH_SIGNATURE_HTML.matcher(content);
+                if (dashSignatureHtml.find()) {
+                    Matcher blockquoteStart = BLOCKQUOTE_START.matcher(content);
+                    Matcher blockquoteEnd = BLOCKQUOTE_END.matcher(content);
+                    List<Integer> start = new ArrayList<>();
+                    List<Integer> end = new ArrayList<>();
+
+                    while (blockquoteStart.find()) {
+                        start.add(blockquoteStart.start());
+                    }
+                    while (blockquoteEnd.find()) {
+                        end.add(blockquoteEnd.start());
+                    }
+                    if (start.size() != end.size()) {
+                        Log.d(K9.LOG_TAG, "There are " + start.size() + " <blockquote> tags, but " +
+                                end.size() + " </blockquote> tags. Refusing to strip.");
+                    } else if (start.size() > 0) {
+                        // Ignore quoted signatures in blockquotes.
+                        dashSignatureHtml.region(0, start.get(0));
+                        if (dashSignatureHtml.find()) {
+                            // before first <blockquote>.
+                            content = content.substring(0, dashSignatureHtml.start());
+                        } else {
+                            for (int i = 0; i < start.size() - 1; i++) {
+                                // within blockquotes.
+                                if (end.get(i) < start.get(i + 1)) {
+                                    dashSignatureHtml.region(end.get(i), start.get(i + 1));
+                                    if (dashSignatureHtml.find()) {
+                                        content = content.substring(0, dashSignatureHtml.start());
+                                        break;
+                                    }
+                                }
+                            }
+                            if (end.get(end.size() - 1) < content.length()) {
+                                // after last </blockquote>.
+                                dashSignatureHtml.region(end.get(end.size() - 1), content.length());
+                                if (dashSignatureHtml.find()) {
+                                    content = content.substring(0, dashSignatureHtml.start());
+                                }
+                            }
+                        }
+                    } else {
+                        // No blockquotes found.
+                        content = content.substring(0, dashSignatureHtml.start());
+                    }
+                }
+
+                // Fix the stripping off of closing tags if a signature was stripped,
+                // as well as clean up the HTML of the quoted message.
+                HtmlCleaner cleaner = new HtmlCleaner();
+                CleanerProperties properties = cleaner.getProperties();
+
+                // see http://htmlcleaner.sourceforge.net/parameters.php for descriptions
+                properties.setNamespacesAware(false);
+                properties.setAdvancedXmlEscape(false);
+                properties.setOmitXmlDeclaration(true);
+                properties.setOmitDoctypeDeclaration(false);
+                properties.setTranslateSpecialEntities(false);
+                properties.setRecognizeUnicodeChars(false);
+
+                TagNode node = cleaner.clean(content);
+                SimpleHtmlSerializer htmlSerialized = new SimpleHtmlSerializer(properties);
+                content = htmlSerialized.getAsString(node, "UTF8");
+            }
+
+            // Add the HTML reply header to the top of the content.
+            quotedHtmlContent = QuotedMessageHelper.quoteOriginalHtmlMessage(
+                    resources, sourceMessage, content, quoteStyle);
+
+            // Load the message with the reply header.
+            view.setQuotedHtml(quotedHtmlContent.getQuotedContent());
+
+            // TODO: Also strip the signature from the text/plain part
+            view.setQuotedText(QuotedMessageHelper.quoteOriginalTextMessage(resources, sourceMessage,
+                    QuotedMessageHelper.getBodyTextFromMessage(sourceMessage, SimpleMessageFormat.TEXT),
+                    quoteStyle, account.getQuotePrefix()));
+
+        } else if (quotedTextFormat == SimpleMessageFormat.TEXT) {
+            if (account.isStripSignature() && (action == Action.REPLY || action == Action.REPLY_ALL)) {
+                if (DASH_SIGNATURE_PLAIN.matcher(content).find()) {
+                    content = DASH_SIGNATURE_PLAIN.matcher(content).replaceFirst("\r\n");
+                }
+            }
+
+            view.setQuotedText(QuotedMessageHelper.quoteOriginalTextMessage(
+                    resources, sourceMessage, content, quoteStyle, account.getQuotePrefix()));
+        }
+
+        if (showQuotedText) {
+            showOrHideQuotedText(QuotedTextMode.SHOW);
+        } else {
+            showOrHideQuotedText(QuotedTextMode.HIDE);
+        }
+    }
+
+    public void builderSetProperties(MessageBuilder builder) {
+        builder.setQuoteStyle(quoteStyle)
+                // TODO avoid using a getter from the view!
+                .setQuotedText(view.getQuotedText())
+                .setQuotedTextMode(quotedTextMode)
+                .setQuotedHtmlContent(quotedHtmlContent)
+                .setReplyAfterQuote(account.isReplyAfterQuote());
+    }
+
+    public void onSaveInstanceState(Bundle outState) {
+        outState.putSerializable(STATE_KEY_QUOTED_TEXT_MODE, quotedTextMode);
+        outState.putSerializable(STATE_KEY_HTML_QUOTE, quotedHtmlContent);
+        outState.putSerializable(STATE_KEY_QUOTED_TEXT_FORMAT, quotedTextFormat);
+        outState.putBoolean(STATE_KEY_FORCE_PLAIN_TEXT, forcePlainText);
+    }
+
+    public void onRestoreInstanceState(Bundle savedInstanceState) {
+        quotedHtmlContent = (InsertableHtmlContent) savedInstanceState.getSerializable(STATE_KEY_HTML_QUOTE);
+        if (quotedHtmlContent != null && quotedHtmlContent.getQuotedContent() != null) {
+            view.setQuotedHtml(quotedHtmlContent.getQuotedContent());
+        }
+        quotedTextFormat = (SimpleMessageFormat) savedInstanceState.getSerializable(
+                STATE_KEY_QUOTED_TEXT_FORMAT);
+        forcePlainText = savedInstanceState.getBoolean(STATE_KEY_FORCE_PLAIN_TEXT);
+
+        showOrHideQuotedText(
+                (QuotedTextMode) savedInstanceState.getSerializable(STATE_KEY_QUOTED_TEXT_MODE));
+    }
+
+    public void processMessageToForward(Message message) throws MessagingException {
+        quoteStyle = QuoteStyle.HEADER;
+        populateUIWithQuotedMessage(message, true, Action.FORWARD);
+    }
+
+    public void initFromReplyToMessage(Message message, Action action)
+            throws MessagingException {
+        populateUIWithQuotedMessage(message, account.isDefaultQuotedTextShown(), action);
+    }
+
+    public void processDraftMessage(LocalMessage message, Map<IdentityField, String> k9identity)
+            throws MessagingException {
+        quoteStyle = k9identity.get(IdentityField.QUOTE_STYLE) != null
+                ? QuoteStyle.valueOf(k9identity.get(IdentityField.QUOTE_STYLE))
+                : account.getQuoteStyle();
+
+        int cursorPosition = 0;
+        if (k9identity.containsKey(IdentityField.CURSOR_POSITION)) {
+            try {
+                cursorPosition = Integer.parseInt(k9identity.get(IdentityField.CURSOR_POSITION));
+            } catch (Exception e) {
+                Log.e(K9.LOG_TAG, "Could not parse cursor position for MessageCompose; continuing.", e);
+            }
+        }
+
+        String showQuotedTextMode;
+        if (k9identity.containsKey(IdentityField.QUOTED_TEXT_MODE)) {
+            showQuotedTextMode = k9identity.get(IdentityField.QUOTED_TEXT_MODE);
+        } else {
+            showQuotedTextMode = "NONE";
+        }
+
+        Integer bodyLength = k9identity.get(IdentityField.LENGTH) != null
+                ? Integer.valueOf(k9identity.get(IdentityField.LENGTH))
+                : 0;
+        Integer bodyOffset = k9identity.get(IdentityField.OFFSET) != null
+                ? Integer.valueOf(k9identity.get(IdentityField.OFFSET))
+                : 0;
+        Integer bodyFooterOffset = k9identity.get(IdentityField.FOOTER_OFFSET) != null
+                ? Integer.valueOf(k9identity.get(IdentityField.FOOTER_OFFSET))
+                : null;
+        Integer bodyPlainLength = k9identity.get(IdentityField.PLAIN_LENGTH) != null
+                ? Integer.valueOf(k9identity.get(IdentityField.PLAIN_LENGTH))
+                : null;
+        Integer bodyPlainOffset = k9identity.get(IdentityField.PLAIN_OFFSET) != null
+                ? Integer.valueOf(k9identity.get(IdentityField.PLAIN_OFFSET))
+                : null;
+
+        QuotedTextMode quotedMode;
+        try {
+            quotedMode = QuotedTextMode.valueOf(showQuotedTextMode);
+        } catch (Exception e) {
+            quotedMode = QuotedTextMode.NONE;
+        }
+
+        // Always respect the user's current composition format preference, even if the
+        // draft was saved in a different format.
+        // TODO - The current implementation doesn't allow a user in HTML mode to edit a draft that wasn't saved with K9mail.
+        String messageFormatString = k9identity.get(IdentityField.MESSAGE_FORMAT);
+
+        MessageFormat messageFormat = null;
+        if (messageFormatString != null) {
+            try {
+                messageFormat = MessageFormat.valueOf(messageFormatString);
+            } catch (Exception e) { /* do nothing */ }
+        }
+
+        if (messageFormat == null) {
+            // This message probably wasn't created by us. The exception is legacy
+            // drafts created before the advent of HTML composition. In those cases,
+            // we'll display the whole message (including the quoted part) in the
+            // composition window. If that's the case, try and convert it to text to
+            // match the behavior in text mode.
+            view.setMessageContentCharacters(QuotedMessageHelper.getBodyTextFromMessage(message, SimpleMessageFormat.TEXT));
+            forcePlainText = true;
+
+            showOrHideQuotedText(quotedMode);
+            return;
+        }
+
+        if (messageFormat == MessageFormat.HTML) {
+            Part part = MimeUtility.findFirstPartByMimeType(message, "text/html");
+            if (part != null) { // Shouldn't happen if we were the one who saved it.
+                quotedTextFormat = SimpleMessageFormat.HTML;
+                String text = MessageExtractor.getTextFromPart(part);
+                if (K9.DEBUG) {
+                    Log.d(K9.LOG_TAG, "Loading message with offset " + bodyOffset + ", length " + bodyLength + ". Text length is " + text.length() + ".");
+                }
+
+                if (bodyOffset + bodyLength > text.length()) {
+                    // The draft was edited outside of K-9 Mail?
+                    Log.d(K9.LOG_TAG, "The identity field from the draft contains an invalid LENGTH/OFFSET");
+                    bodyOffset = 0;
+                    bodyLength = 0;
+                }
+                // Grab our reply text.
+                String bodyText = text.substring(bodyOffset, bodyOffset + bodyLength);
+                view.setMessageContentCharacters(HtmlConverter.htmlToText(bodyText));
+
+                // Regenerate the quoted html without our user content in it.
+                StringBuilder quotedHTML = new StringBuilder();
+                quotedHTML.append(text.substring(0, bodyOffset));   // stuff before the reply
+                quotedHTML.append(text.substring(bodyOffset + bodyLength));
+                if (quotedHTML.length() > 0) {
+                    quotedHtmlContent = new InsertableHtmlContent();
+                    quotedHtmlContent.setQuotedContent(quotedHTML);
+                    // We don't know if bodyOffset refers to the header or to the footer
+                    quotedHtmlContent.setHeaderInsertionPoint(bodyOffset);
+                    if (bodyFooterOffset != null) {
+                        quotedHtmlContent.setFooterInsertionPoint(bodyFooterOffset);
+                    } else {
+                        quotedHtmlContent.setFooterInsertionPoint(bodyOffset);
+                    }
+                    view.setQuotedHtml(quotedHtmlContent.getQuotedContent());
+                }
+            }
+            if (bodyPlainOffset != null && bodyPlainLength != null) {
+                processSourceMessageText(message, bodyPlainOffset, bodyPlainLength, false);
+            }
+        } else if (messageFormat == MessageFormat.TEXT) {
+            quotedTextFormat = SimpleMessageFormat.TEXT;
+            processSourceMessageText(message, bodyOffset, bodyLength, true);
+        } else {
+            Log.e(K9.LOG_TAG, "Unhandled message format.");
+        }
+
+        // Set the cursor position if we have it.
+        try {
+            view.setMessageContentCursorPosition(cursorPosition);
+        } catch (Exception e) {
+            Log.e(K9.LOG_TAG, "Could not set cursor position in MessageCompose; ignoring.", e);
+        }
+
+        showOrHideQuotedText(quotedMode);
+    }
+
+    /**
+     * Pull out the parts of the now loaded source message and apply them to the new message
+     * depending on the type of message being composed.
+     * @param message Source message
+     * @param bodyOffset Insertion point for reply.
+     * @param bodyLength Length of reply.
+     * @param viewMessageContent Update mMessageContentView or not.
+     * @throws MessagingException
+     */
+    private void processSourceMessageText(Message message, Integer bodyOffset, Integer bodyLength,
+            boolean viewMessageContent) throws MessagingException {
+        Part textPart = MimeUtility.findFirstPartByMimeType(message, "text/plain");
+        if (textPart != null) {
+            String text = MessageExtractor.getTextFromPart(textPart);
+            if (K9.DEBUG) {
+                Log.d(K9.LOG_TAG, "Loading message with offset " + bodyOffset + ", length " + bodyLength + ". Text length is " + text.length() + ".");
+            }
+
+            // If we had a body length (and it was valid), separate the composition from the quoted text
+            // and put them in their respective places in the UI.
+            if (bodyLength > 0) {
+                try {
+                    String bodyText = text.substring(bodyOffset, bodyOffset + bodyLength);
+
+                    // Regenerate the quoted text without our user content in it nor added newlines.
+                    StringBuilder quotedText = new StringBuilder();
+                    if (bodyOffset == 0 && text.substring(bodyLength, bodyLength + 4).equals("\r\n\r\n")) {
+                        // top-posting: ignore two newlines at start of quote
+                        quotedText.append(text.substring(bodyLength + 4));
+                    } else if (bodyOffset + bodyLength == text.length() &&
+                            text.substring(bodyOffset - 2, bodyOffset).equals("\r\n")) {
+                        // bottom-posting: ignore newline at end of quote
+                        quotedText.append(text.substring(0, bodyOffset - 2));
+                    } else {
+                        quotedText.append(text.substring(0, bodyOffset));   // stuff before the reply
+                        quotedText.append(text.substring(bodyOffset + bodyLength));
+                    }
+
+                    if (viewMessageContent) {
+                        view.setMessageContentCharacters(bodyText);
+                    }
+
+                    view.setQuotedText(quotedText.toString());
+                } catch (IndexOutOfBoundsException e) {
+                    // Invalid bodyOffset or bodyLength.  The draft was edited outside of K-9 Mail?
+                    Log.d(K9.LOG_TAG, "The identity field from the draft contains an invalid bodyOffset/bodyLength");
+                    if (viewMessageContent) {
+                        view.setMessageContentCharacters(text);
+                    }
+                }
+            } else {
+                if (viewMessageContent) {
+                    view.setMessageContentCharacters(text);
+                }
+            }
+        }
+    }
+
+    public void onClickShowQuotedText() {
+        showOrHideQuotedText(QuotedTextMode.SHOW);
+        messageCompose.updateMessageFormat();
+        messageCompose.saveDraftEventually();
+    }
+
+    public void onClickDeleteQuotedText() {
+        showOrHideQuotedText(QuotedTextMode.HIDE);
+        messageCompose.updateMessageFormat();
+        messageCompose.saveDraftEventually();
+    }
+
+    public void onClickEditQuotedText() {
+        forcePlainText = true;
+        messageCompose.loadQuotedTextForEdit();
+    }
+
+    public boolean includeQuotedText() {
+        return quotedTextMode == QuotedTextMode.SHOW;
+    }
+
+    public boolean isForcePlainText() {
+        return forcePlainText;
+    }
+
+    public boolean isQuotedTextText() {
+        return quotedTextFormat == SimpleMessageFormat.TEXT;
+    }
+
+}
